Update cloud-functions.md
While we know that some servers may take hours to spin up I fear that a lot of people reading this doc will not have experienced this problem and will find the statement hyperbolic. That will lead them to be skeptical about other parts of the doc.
I’d say just change it to such applications run on servers that take minutes to spin up which still gets the point across.